feat: add editor drawer component
