+++ src/components/Drawer/Drawer.jsx
@@ -1,127 +1,55 @@
-
-
 import React from 'react';
-import { Drawer as AntDrawer, Button, Switch, InputNumber, Select } from 'antd';
+import { Drawer as AntDrawer, Button } from 'antd';
 import { MenuOutlined } from '@ant-design/icons';
 
 /**
- * EditorDrawer
- * Drawer UI for entry editing controls. Pure presentational; all state/handlers
- * are passed from parent.
+ * Minimal Drawer component that provides structural slots for custom content.
  */
-export default function EditorDrawer({
-  // visibility + layout
-  drawerOpen,
-  drawerWidth = 300,
+export default function Drawer({
+  open,
+  width = 300,
   onHamburgerClick,
   onMouseEnter,
   onMouseLeave,
-
-  // editor controls
-  pomodoroEnabled,
-  onPomodoroToggle,
-  maxWidth,
-  onMaxWidthChange,
-
-  // entry context
-  type,
-  mode,
-  aliases,
-  groups = [],
-  selectedSubgroupId,
-  onChangeSubgroup, // (val) => void
-
-  // actions
-  onSave,
-  onDelete,
-  onArchive,
-  onCancel,
-
-  // keyboard help
-  showShortcutList,
-  onToggleShortcutList,
-  entryShortcuts = [],
+  header,
+  body,
+  footer,
+  children,
+  ...props
 }) {
-  const subgroupOptions = groups.flatMap((g) =>
-    g.subgroups.map((s) => ({ value: s.id, label: `${g.name} / ${s.name}` }))
-  );
-
   return (
     <>
-      <Button
-        type="text"
-        icon={<MenuOutlined />}
-        onClick={onHamburgerClick}
-        style={{ position: 'fixed', top: '1rem', right: '1rem', zIndex: 1002 }}
-      />
+      {onHamburgerClick && (
+        <Button
+          type="text"
+          icon={<MenuOutlined />}
+          onClick={onHamburgerClick}
+          style={{ position: 'fixed', top: '1rem', right: '1rem', zIndex: 1002 }}
+        />
+      )}
       <div
         onMouseEnter={onMouseEnter}
         onMouseLeave={onMouseLeave}
-        className="entry-editor-drawer-wrapper"
-        style={{ width: drawerWidth }}
+        className="drawer-wrapper"
+        style={{ width }}
       >
         <AntDrawer
           placement="right"
-          open={drawerOpen}
+          open={open}
           mask={false}
           closable={false}
-          width={drawerWidth}
+          width={width}
           getContainer={false}
           rootStyle={{ position: 'absolute' }}
           body={{ padding: '1rem' }}
+          {...props}
         >
-          <h2 style={{ marginTop: 0 }}>
-            {mode === 'edit' ? `Edit ${aliases.entry}` : `New ${aliases.entry}`}
-          </h2>
-          <div style={{ display: 'flex', flexDirection: 'column', gap: '0.5rem' }}>
-            <div style={{ display: 'flex', alignItems: 'center', gap: '0.25rem' }}>
-              <span>Pomodoro</span>
-              <Switch checked={pomodoroEnabled} onChange={onPomodoroToggle} size="small" />
-            </div>
-            <div style={{ display: 'flex', alignItems: 'center', gap: '0.25rem' }}>
-              <span>Max Width</span>
-              <InputNumber
-                min={25}
-                max={95}
-                step={1}
-                value={maxWidth}
-                onChange={onMaxWidthChange}
-                size="small"
-                formatter={(value) => `${value}%`}
-                parser={(value) => value.replace('%', '')}
-              />
-            </div>
-            {type === 'entry' && groups.length > 0 && (
-              <Select
-                value={selectedSubgroupId}
-                onChange={onChangeSubgroup}
-                options={subgroupOptions}
-                size="small"
-              />
-            )}
-            <Button className="drawer-btn drawer-btn-save" onClick={onSave}>Save</Button>
-            {mode === 'edit' && onDelete && (
-              <Button className="drawer-btn drawer-btn-delete" onClick={onDelete}>Delete</Button>
-            )}
-            {mode === 'edit' && onArchive && (
-              <Button className="drawer-btn drawer-btn-archive" onClick={onArchive}>
-                Archive/Restore
-              </Button>
-            )}
-            <Button className="drawer-btn drawer-btn-cancel" onClick={onCancel}>Cancel</Button>
-            <Button type="link" onClick={onToggleShortcutList}>Keyboard Shortcuts</Button>
-            {showShortcutList && (
-              <ul style={{ paddingLeft: '1rem' }}>
-                {entryShortcuts.map((s) => (
-                  <li key={s.action}>
-                    {s.action}: {s.keys}
-                  </li>
-                ))}
-              </ul>
-            )}
-          </div>
+          {header}
+          {body || children}
+          {footer}
         </AntDrawer>
       </div>
     </>
   );
 }
+
